requirements.txt
# Herramientas de desarrollo
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
pylint==2.17.5

# Seguridad
bandit==1.7.5
safety==2.3.5

# Servidor web
flask==2.3.3
gunicorn==21.2.0

# Utilidades
pyyaml==6.0.1
requests==2.31.0